refactor(worklist): the lanes split by concept, as the gate asked
Three new lanes pushed both the feed and its seam past the file and function
ceilings. The gate was right about the cause: each lane had been written out in
full, and the fourth copy of a fifteen-line shape is duplication rather than
length.

The optional lanes — meetings, deal risk, promises — now live in their own file
with the shape they share. Bound or absent, read, render, and either fill the
lane or name it withheld. The required lanes keep their own switch in one
helper for the same reason.

The seam's three adapters move to a file of their own. Each is a binding rather
than an implementation: promises from the people module's claim read, deal risk
from the same engine whats_slipping_this_week reads, meetings from the
activities list. That is what the file now says out loud.

No behaviour moved with them. The same tests pass, and the browser draws the
same three lanes.
